+++ keramik/src/ipfs.md
@@ -1,10 +1,12 @@
 # IPFS
 
-The IPFS behavior used by Ceramic can be customized.
+The IPFS behavior used by CAS and Ceramic can be customized using the same IPFS spec.
 
 ## Rust IPFS
 
-Example [network config](./setup_network.md) that uses Rust based IPFS (i.e. ceramic-one) with its defaults.
+### Ceramic
+
+Example [network config](./setup_network.md) that uses Rust based IPFS (i.e. ceramic-one) with its defaults for Ceramic.
 
 ```yaml
 apiVersion: "keramik.3box.io/v1alpha1"
@@ -18,7 +20,7 @@ spec:
         rust: {}
 ```
 
-Example [network config](./setup_network.md) that uses Rust based IPFS (i.e. ceramic-one) with a specific image.
+Example [network config](./setup_network.md) that uses Rust based IPFS (i.e. ceramic-one) with a specific image for Ceramic.
 
 ```yaml
 apiVersion: "keramik.3box.io/v1alpha1"
@@ -34,9 +36,43 @@ spec:
          imagePullPolicy: IfNotPresent
 ```
 
+### CAS
+
+Example [network config](./setup_network.md) that uses Rust based IPFS (i.e. ceramic-one) with a specific image for CAS.
+
+```yaml
+apiVersion: "keramik.3box.io/v1alpha1"
+kind: Network
+metadata:
+  name: example-vanilla-ceramic-one
+spec:
+  replicas: 5
+  cas:
+    ipfs:
+      rust: {}
+```
+
+Example [network config](./setup_network.md) that uses Rust based IPFS (i.e. ceramic-one) with a specific image for CAS.
+
+```yaml
+apiVersion: "keramik.3box.io/v1alpha1"
+kind: Network
+metadata:
+  name: example-custom-ceramic-one
+spec:
+  replicas: 5
+  cas:
+    ipfs:
+     rust:
+       image: rust-ceramic/ceramic-one:dev
+       imagePullPolicy: IfNotPresent
+```
+
 ## Kubo IPFS
 
-Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with its defaults.
+### Ceramic
+
+Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with its defaults for Ceramic.
 
 ```yaml
 apiVersion: "keramik.3box.io/v1alpha1"
@@ -50,7 +86,7 @@ spec:
         go: {}
 ```
 
-Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with a specific image.
+Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with a specific image for Ceramic.
 
 ```yaml
 apiVersion: "keramik.3box.io/v1alpha1"
@@ -66,7 +102,7 @@ spec:
          imagePullPolicy: IfNotPresent
 ```
 
-Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with extra configuration commands.
+Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with extra configuration commands for Ceramic.
 
 ```yaml
 apiVersion: "keramik.3box.io/v1alpha1"
@@ -82,4 +118,54 @@ spec:
          imagePullPolicy: IfNotPresent
          commands:
            - ipfs config --json Swarm.RelayClient.Enabled false
+```
+
+### CAS
+
+Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with its defaults for CAS.
+
+```yaml
+apiVersion: "keramik.3box.io/v1alpha1"
+kind: Network
+metadata:
+  name: example-vanilla-kubo
+spec:
+  replicas: 5
+  cas:
+    ipfs:
+      go: {}
+```
+
+Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with a specific image for CAS.
+
+```yaml
+apiVersion: "keramik.3box.io/v1alpha1"
+kind: Network
+metadata:
+  name: example-custom-kubo
+spec:
+  replicas: 5
+  cas:
+    ipfs:
+     go:
+       image: ceramicnetwork/go-ipfs-daemon:develop
+       imagePullPolicy: IfNotPresent
+```
+
+Example [network config](./setup_network.md) that uses Go based IPFS (i.e. Kubo) with extra configuration commands for CAS.
+
+```yaml
+apiVersion: "keramik.3box.io/v1alpha1"
+kind: Network
+metadata:
+  name: example-custom-kubo
+spec:
+  replicas: 5
+  cas:
+    ipfs:
+     go:
+       image: ceramicnetwork/go-ipfs-daemon:develop
+       imagePullPolicy: IfNotPresent
+       commands:
+         - ipfs config --json Swarm.RelayClient.Enabled false
 ```
